requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
algoliasearch==4.11.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.9.0
bottle==0.13.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.7
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.16.1
fire==0.5.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
fuzzywuzzy==0.18.0
gradio==4.31.2
gradio_client==0.16.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
imagecodecs==2024.6.1
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib_resources==6.4.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.4
Levenshtein==0.25.1
linkify-it-py==2.0.3
llvmlite==0.43.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdit-py-plugins==0.4.1
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.60.0
numcodecs==0.13.0
numpy==1.26.4
onnxruntime==1.19.2
openai==1.46.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.4
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
pexpect==4.9.0
phpserialize==1.3
pillow==11.0.0
pillow_heif==0.20.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
prompt-toolkit==3.0.36
propcache==0.2.1
protobuf==5.28.3
proxy-tools==0.1.0
psutil==6.0.0
ptyprocess==0.7.0
pycaw==20240210
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyjxl==0.1.0
PyMatting==1.1.12
PyNaCl==1.5.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyreadline3==3.5.4
PyRect==0.2.0
PySocks==1.7.1
pystray==0.19.5
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-Levenshtein==0.25.1
python-multipart==0.0.9
pythonnet==3.0.4
pytidylib==0.3.2
pytube==15.0.0
pytz==2024.1
pywebview==5.3
pywin32==306
PyYAML==6.0.1
questionary==2.0.1
rapidfuzz==3.9.5
referencing==0.35.1
regex==2024.5.10
rembg==2.0.59
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.4
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.25.0
semantic-version==2.10.0
sentence-transformers==3.0.1
shellingham==1.5.4
simple-lama-inpainting==0.1.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
sympy==1.13.3
tbb==2021.12.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.7.21
tkhtmlview==0.3.1
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch @ file:///C:/Users/rx/torch-2.4.1%2Bcu118-cp310-cp310-win_amd64.whl
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.66.4
transformers==4.40.2
transliterate==1.10.2
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.3
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.29.0
vboxapi==1.0
watchfiles==0.21.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wexpect==4.0.0
whisper==1.0
wsproto==1.2.0
xlwt==1.3.0
yarl==1.18.3
